feat:Mise à jour des vues et amélioration de la gestion des utilisateurs

+++ projects/views.py
@@ -1,7 +1,9 @@
+from pyexpat.errors import messages
+
+from django.http import HttpResponseForbidden
 from django.shortcuts import get_object_or_404, render, redirect
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
-
 import projects
 from .models import Projet, Tache
 
@@ -26,6 +28,8 @@ def projet_create(request):
             description=request.POST.get("description"),
             createur=request.user
         )
+          messages.success(request, "Le projet a été créé avec succès.")
+
           
           return redirect("dashboard")
 
@@ -49,11 +53,16 @@ def projet_detail(request, id):
 
 @login_required
 def projet_delete(request, id):
-   projet = get_object_or_404(Projet, id=id)    
-   projet.delete()
+    projet = get_object_or_404(Projet, id=id)
 
-   return redirect("dashboard")
+    if projet.createur != request.user:
+        return HttpResponseForbidden(
+            "Vous n'avez pas l'autorisation de supprimer ce projet."
+        )
 
+    projet.delete()
+    messages.success(request, "Le projet a été supprimé avec succès.")
+    return redirect("dashboard")
 
 # vues projet_update
 
@@ -91,6 +100,7 @@ def tache_create(request):
             priorite=request.POST.get("priorite"),
             deadline=request.POST.get("deadline"),
         )
+        messages.success(request, "La tâche a été créée avec succès.")
 
         return redirect("dashboard")
